backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.0
gunicorn==23.0.0
psycopg[binary]==3.2.9
redis==6.0.0
whitenoise==6.9.0
